Unify /dashboard and /org/repo pages
To make sure that the pages load data and handle errors in the same way
