+++ src/lib/components/dashboard/index.ts
@@ -2,3 +2,7 @@ export { default as Jobs } from './Jobs.svelte';
 export { default as Stats } from './Stats.svelte';
 export { default as BeforeAfter } from './BeforeAfter.svelte';
 export { default as SubscribeButton } from './SubscripeButton.svelte';
+export { default as Dashboard } from './Dashboard.svelte';
+export { default as OrganisationPicker } from './OrganisationPicker.svelte';
+export { default as All } from './All.svelte';
+export { default as Repository } from './Repository.svelte';
