+++ src/routes/[organization]/[repository].svelte
@@ -8,84 +8,10 @@
 </script>
 
 <script lang="ts">
-	import { isPublic as isJobPublic, list, type Job } from '$lib/jobs';
-	import { Jobs, Stats } from '$lib/components/dashboard';
-	import { Subscribe } from '$lib/components/subscriptions';
+	import { Dashboard } from '$lib/components/dashboard';
 	import { page } from '$app/stores';
-	import { GitHubLoginButton } from '$lib/components/index';
-	import { onMount } from 'svelte';
-	import Spinner from '$lib/Spinner.svelte';
-	import { compareAsc } from 'date-fns';
-
 	export let login: string | null;
-
 	const { organization, repository } = $page.params;
-
-	let jobs: Job[] = [];
-	let loaded = false;
-	let loading = false;
-
-	const loadJobs = async () => {
-		let cursor = '';
-		loading = true;
-		for (let i = 0; i < 10; i++) {
-			let stop = false;
-
-			await list({ organization, repository, cursor, onlyRootJobs: true }).then((data) => {
-				jobs = jobs.concat(data.jobs);
-				cursor = data.next ?? '';
-				loaded = true;
-				if (!data.next) {
-					stop = true;
-				}
-			});
-
-			if (stop) {
-				break;
-			}
-		}
-
-		loading = false;
-	};
-
-	$: latestJob = jobs
-		.sort((a, b) => compareAsc(new Date(a.created_at), new Date(b.created_at)))
-		.slice(-1)[0];
-	$: isPublic = latestJob ? isJobPublic(latestJob) : false;
-
-	onMount(loadJobs);
 </script>
 
-<div class="flex flex-col space-y-2 font-mono">
-	{#if !login}
-		<div class="mt-8 flex flex-col items-center space-y-2 text-gray-600">
-			<p>You need to login to access this page</p>
-			<GitHubLoginButton {login} />
-		</div>
-	{:else if !loaded}
-		<div class="flex justify-around">
-			<Spinner />
-		</div>
-	{:else if jobs?.length > 0}
-		{#if loading}
-			<div class="flex justify-around">
-				<Spinner />
-			</div>
-		{/if}
-
-		{#if isPublic}
-			<h2>Thank you for being open source!</h2>
-			<p>Codeball is free for open source projects.</p>
-		{:else}
-			<Subscribe {organization} />
-		{/if}
-
-		<Stats {jobs} />
-		<!-- <BeforeAfter {jobs} /> -->
-		<Jobs {jobs} />
-	{:else}
-		<div class="mt-8 flex flex-col items-center space-y-2 text-gray-600">
-			<p>Something went wrong, or no jobs found? Please come back later.</p>
-		</div>
-	{/if}
-</div>
+<Dashboard {organization} {repository} {login} />
